Move python files outside the root of the project
