+++ package.json
@@ -6,7 +6,7 @@
     "build-css": "sass static/sass/styles.scss static/css/styles.css --load-path=node_modules --style=compressed && postcss --map false --use autoprefixer --replace 'static/css/**/*.css'",
     "serve": "./entrypoint 0.0.0.0:${PORT}",
     "start": "yarn run build && yarn run serve",
-    "migrate": "python3 main.py"
+    "migrate": "python3 webapp/migrate.py"
   },
   "devDependencies": {
     "autoprefixer": "10.2.4",
